Commit#5

+++ src/Character.java
@@ -7,7 +7,7 @@ public abstract class Character {
     byte strength;
     byte experience;
     byte damage;
-    byte meds;
+    int meds;
 
     public Character() {
         this.name = null;
